+++ src/dma/traits.rs
@@ -1,6 +1,7 @@
 use super::*;
 use crate::{
     adc::Adc,
+    i2c,
     pac::{self, DMA1, DMA2},
     serial, spi, timer,
 };
@@ -333,6 +334,7 @@ dma_map!(
     (Stream2<DMA1>, 0, pac::SPI3, PeripheralToMemory),       //SPI3_RX
     (Stream2<DMA1>, 0, spi::Rx<pac::SPI3>, PeripheralToMemory), //SPI3_RX
     (Stream4<DMA1>, 3, pac::I2C3, MemoryToPeripheral),       //I2C3_TX
+    (Stream4<DMA1>, 3, i2c::Tx<pac::I2C3>, MemoryToPeripheral),       //I2C3_TX
     (Stream5<DMA1>, 0, pac::SPI3, MemoryToPeripheral),       //SPI3_TX
     (Stream5<DMA1>, 0, spi::Tx<pac::SPI3>, MemoryToPeripheral), //SPI3_TX
     (Stream7<DMA1>, 0, pac::SPI3, MemoryToPeripheral),       //SPI3_TX
@@ -430,19 +432,24 @@ dma_map!(
     (Stream6<DMA2>, 0, timer::CCR3<pac::TIM1>, PeripheralToMemory), //TIM1_CH3
     (Stream6<DMA2>, 6, timer::CCR3<pac::TIM1>, MemoryToPeripheral), //TIM1_CH3
     (Stream6<DMA2>, 6, timer::CCR3<pac::TIM1>, PeripheralToMemory), //TIM1_CH3
-    (Stream0<DMA1>, 1, pac::I2C1, PeripheralToMemory),       //I2C1_RX
-    (Stream2<DMA1>, 7, pac::I2C2, PeripheralToMemory),       //I2C2_RX
+    (Stream0<DMA1>, 1, pac::I2C1, PeripheralToMemory),              //I2C1_RX
+    (Stream0<DMA1>, 1, i2c::Rx<pac::I2C1>, PeripheralToMemory),     //I2C1_RX
+    (Stream2<DMA1>, 7, pac::I2C2, PeripheralToMemory),              //I2C2_RX
+    (Stream2<DMA1>, 7, i2c::Rx<pac::I2C2>, PeripheralToMemory),     //I2C2_RX
     (Stream3<DMA1>, 0, pac::SPI2, PeripheralToMemory),       //SPI2_RX
     (Stream3<DMA1>, 0, spi::Rx<pac::SPI2>, PeripheralToMemory), //SPI2_RX
-    (Stream3<DMA1>, 7, pac::I2C2, PeripheralToMemory),       //I2C2_RX
+    (Stream3<DMA1>, 7, pac::I2C2, PeripheralToMemory),          //I2C2_RX
+    (Stream3<DMA1>, 7, spi::Rx<pac::I2C2>, PeripheralToMemory), //I2C2_RX
     (Stream4<DMA1>, 0, pac::SPI2, MemoryToPeripheral),       //SPI2_TX
     (Stream4<DMA1>, 0, spi::Tx<pac::SPI2>, MemoryToPeripheral), // SPI2_TX
-    (Stream5<DMA1>, 1, pac::I2C1, PeripheralToMemory),       //I2C1_RX
+    (Stream5<DMA1>, 1, pac::I2C1, PeripheralToMemory),          //I2C1_RX
+    (Stream5<DMA1>, 1, i2c::Rx<pac::I2C1>, PeripheralToMemory), //I2C1_RX
     (Stream5<DMA1>, 4, pac::USART2, PeripheralToMemory),     //USART2_RX
     (Stream5<DMA1>, 4, serial::Rx<pac::USART2>, PeripheralToMemory), //USART2_RX
     (Stream6<DMA1>, 4, pac::USART2, MemoryToPeripheral),     //USART2_TX
     (Stream6<DMA1>, 4, serial::Tx<pac::USART2>, MemoryToPeripheral), //USART2_TX
-    (Stream7<DMA1>, 7, pac::I2C2, MemoryToPeripheral),       //I2C2_TX
+    (Stream7<DMA1>, 7, pac::I2C2, MemoryToPeripheral),          //I2C2_TX
+    (Stream7<DMA1>, 7, i2c::Tx<pac::I2C2>, MemoryToPeripheral), //I2C2_TX
     (Stream0<DMA2>, 0, pac::ADC1, PeripheralToMemory),       //ADC1
     (Stream0<DMA2>, 0, Adc<pac::ADC1>, PeripheralToMemory),
     (Stream0<DMA2>, 3, pac::SPI1, PeripheralToMemory), //SPI1_RX
@@ -529,8 +536,10 @@ address!(
     feature = "stm32f446",
 ))]
 dma_map!(
-    (Stream1<DMA1>, 1, pac::I2C3, PeripheralToMemory), //I2C3_RX
-    (Stream2<DMA1>, 3, pac::I2C3, PeripheralToMemory), //I2C3_RX:DMA_CHANNEL_3
+    (Stream1<DMA1>, 1, pac::I2C3, PeripheralToMemory),          //I2C3_RX
+    (Stream1<DMA1>, 1, i2c::Rx<pac::I2C3>, PeripheralToMemory), //I2C3_RX
+    (Stream2<DMA1>, 3, pac::I2C3, PeripheralToMemory),          //I2C3_RX:DMA_CHANNEL_3
+    (Stream2<DMA1>, 3, i2c::Rx<pac::I2C3>, PeripheralToMemory), //I2C3_RX:DMA_CHANNEL_3
 );
 
 #[cfg(any(feature = "stm32f401", feature = "stm32f411",))]
@@ -552,7 +561,10 @@ dma_map!(
     feature = "stm32f413",
     feature = "stm32f423",
 ))]
-dma_map!((Stream5<DMA1>, 6, pac::I2C3, MemoryToPeripheral),); //I2C3_TX:DMA_CHANNEL_6);
+dma_map!(
+    (Stream5<DMA1>, 6, pac::I2C3, MemoryToPeripheral), //I2C3_TX:DMA_CHANNEL_6);
+    (Stream5<DMA1>, 6, i2c::Tx<pac::I2C3>, MemoryToPeripheral), //I2C3_TX:DMA_CHANNEL_6);
+);
 
 #[cfg(any(
     feature = "stm32f401",
@@ -569,8 +581,10 @@ dma_map!((Stream5<DMA1>, 6, pac::I2C3, MemoryToPeripheral),); //I2C3_TX:DMA_CHAN
     feature = "stm32f479",
 ))]
 dma_map!(
-    (Stream6<DMA1>, 1, pac::I2C1, MemoryToPeripheral), //I2C1_TX
-    (Stream7<DMA1>, 1, pac::I2C1, MemoryToPeripheral), //I2C1_TX
+    (Stream6<DMA1>, 1, pac::I2C1, MemoryToPeripheral),          //I2C1_TX
+    (Stream6<DMA1>, 1, i2c::Tx<pac::I2C1>, MemoryToPeripheral), //I2C1_TX
+    (Stream7<DMA1>, 1, pac::I2C1, MemoryToPeripheral),          //I2C1_TX
+    (Stream7<DMA1>, 1, i2c::Tx<pac::I2C1>, MemoryToPeripheral), //I2C1_TX
     (Stream3<DMA2>, 3, pac::SPI1, MemoryToPeripheral), //SPI1_TX
     (Stream3<DMA2>, 3, spi::Tx<pac::SPI1>, MemoryToPeripheral), //SPI1_TX
     (Stream5<DMA2>, 3, pac::SPI1, MemoryToPeripheral), //SPI1_TX
@@ -773,7 +787,8 @@ dma_map!(
     feature = "stm32f479",
 ))]
 dma_map!(
-    (Stream2<DMA1>, 3, pac::I2C3, PeripheralToMemory), //I2C3_RX
+    (Stream2<DMA1>, 3, pac::I2C3, PeripheralToMemory),          //I2C3_RX
+    (Stream2<DMA1>, 3, i2c::Rx<pac::I2C3>, PeripheralToMemory), //I2C3_RX
     (Stream5<DMA2>, 2, pac::CRYP, PeripheralToMemory), //CRYP_OUT
     (Stream6<DMA2>, 2, pac::CRYP, MemoryToPeripheral), //CRYP_IN
     (Stream7<DMA2>, 2, pac::HASH, MemoryToPeripheral), //HASH_IN
@@ -913,9 +928,12 @@ address!(
     feature = "stm32f423",
 ))]
 dma_map!(
-    (Stream1<DMA1>, 0, pac::I2C1, MemoryToPeripheral), //I2C1_TX
-    (Stream6<DMA1>, 1, pac::I2C1, MemoryToPeripheral), //I2C1_TX:DMA_CHANNEL_1
-    (Stream7<DMA1>, 1, pac::I2C1, MemoryToPeripheral), //I2C1_TX:DMA_CHANNEL_1
+    (Stream1<DMA1>, 0, pac::I2C1, MemoryToPeripheral),          //I2C1_TX
+    (Stream1<DMA1>, 0, i2c::Tx<pac::I2C1>, MemoryToPeripheral), //I2C1_TX
+    (Stream6<DMA1>, 1, pac::I2C1, MemoryToPeripheral),          //I2C1_TX:DMA_CHANNEL_1
+    (Stream6<DMA1>, 1, i2c::Tx<pac::I2C1>, MemoryToPeripheral), //I2C1_TX:DMA_CHANNEL_1
+    (Stream7<DMA1>, 1, pac::I2C1, MemoryToPeripheral),          //I2C1_TX:DMA_CHANNEL_1
+    (Stream7<DMA1>, 1, i2c::Tx<pac::I2C1>, MemoryToPeripheral), //I2C1_TX:DMA_CHANNEL_1
     (Stream7<DMA1>, 6, pac::USART2, PeripheralToMemory), //USART2_RX:DMA_CHANNEL_6
     (Stream2<DMA2>, 2, pac::SPI1, MemoryToPeripheral), //SPI1_TX
     (Stream2<DMA2>, 2, spi::Tx<pac::SPI1>, MemoryToPeripheral), //SPI1_TX